bug fix for proxy
