+++ src/js/modules/proxyForWebsite/index.js
@@ -4,6 +4,7 @@
  * Description:
  */
 import {Feature} from 'Libs/feature';
+import _ from 'lodash';
 import {fetchJSON, fetchImage, fetchPOST} from './fetch';
 import {connect} from './connect';
 import {cookie} from './cookie';
@@ -22,6 +23,26 @@ export class ProxyForWebsite extends Feature {
     }
 
     addListener = () => {
+        chrome.webRequest.onBeforeSendHeaders.addListener(details => {
+            const {initiator, requestHeaders, url} = details;
+            const fromHelper = !_.isEmpty(_.find(requestHeaders, ({name, value}) => name === 'From' && value === 'bilibili-helper'));
+            if ((/^chrome-extension:\/\//.test(initiator) || fromHelper) && url.match('/reply/add')) {
+                const originHeader = requestHeaders.find((h) => h.name.toLowerCase() === 'origin');
+                if (originHeader) {
+                    originHeader.value = 'https://h.bilibili.com';
+                } else {
+                    requestHeaders.push({
+                        name: 'Origin',
+                        value: 'https://h.bilibili.com',
+                    });
+                }
+                return {requestHeaders};
+            }
+        }, {
+            urls: [
+                'https://api.bilibili.com/x/v2/reply/add?*',
+            ],
+        }, ['blocking', 'requestHeaders', 'extraHeaders']);
         chrome.runtime.onConnect.addListener((port) => {
             port.onMessage.addListener((message, websitePort) => {
                 const {command, data} = message;
